+++ README.md
@@ -1,0 +1,12 @@
+
+
+A principled approach to data analysis includes data alignment, canonical form and constructing all automorphims.
+As such, each directory contains the following structure.
+
+```bash
+ complete - a complete invariant applied to the data
+ canon - a canonical form of the data
+ dfa - a deterministic finite automaton
+ automorph - all data automorphisms of the group
+ alignment - disentangle the group action from the data
+``
