+++ examples/expl_frame/03_md17.py
@@ -24,8 +24,7 @@ from torch_geometric.loader import DataLoader
 
 from pointgroup import PointGroup
 
-sys.path.append('./torch_canon/')
-from E3Global.CategoricalPointCloud import CatFrame as Frame
+from torch_canon.pointcloud import CanonEn as Canon
 
 
 # Setup
@@ -40,7 +39,7 @@ logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(message)s')
 start_time = time()
 
 md17 = MD17(root='./data/md17/', name=args.name)
-frame = Frame(tol=1e-2, save='all')
+frame = Canon(tol=1e-2, save='all')
 
 atomic_number_to_symbol = {
     1: 'H', 6: 'C', 7: 'N', 8: 'O', 9: 'F'
